+++ src/app/modules/select-columns/select-columns.component.ts
@@ -1,6 +1,8 @@
-import { Component, Inject, OnInit } from '@angular/core';
+import { Component, Inject } from '@angular/core';
 import { ColumnDescription } from '../../types/column.description';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 /**
  * Component for selecting columns displayed in resource table
@@ -10,9 +12,15 @@ import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
   templateUrl: './select-columns.component.html',
   styleUrls: ['./select-columns.component.less']
 })
-export class SelectColumnsComponent implements OnInit {
+export class SelectColumnsComponent {
   resourceType: string;
+  // Array of available column descriptions for the resource table.
   columns: ColumnDescription[] = [];
+  // Observable of the column names of the table columns with data
+  columnsWithData: BehaviorSubject<string[]>;
+  // Observable of the column descriptions of the available table
+  // columns with data
+  columns$: Observable<ColumnDescription[]>;
 
   constructor(
     private dialogRef: MatDialogRef<SelectColumnsComponent>,
@@ -20,20 +28,52 @@ export class SelectColumnsComponent implements OnInit {
   ) {
     this.resourceType = data.resourceType;
     this.columns = data.columns;
+    this.columnsWithData = data.columnsWithData;
+    this.columns$ = this.columnsWithData.pipe(
+      map((columnsWithData) =>
+        this.columns.filter(
+          (column) =>
+            columnsWithData.length === 0 ||
+            columnsWithData.indexOf(column.element) !== -1
+        )
+      )
+    );
   }
 
-  ngOnInit(): void {}
-
+  /**
+   * Closes dialog.
+   */
   close(): void {
     this.dialogRef.close();
   }
 
+  /**
+   * Save columns visibility changes.
+   */
   save(): void {
+    if (!this.hasVisibleSelectedColumns()) {
+      this.clearSelection();
+    }
     this.dialogRef.close({
       columns: this.columns
     });
   }
 
+  /**
+   * Checks if there are some selected columns in the dialog.
+   */
+  hasVisibleSelectedColumns(): boolean {
+    return this.columns.some(
+      (column) =>
+        column.visible &&
+        (this.columnsWithData.value.length === 0 ||
+          this.columnsWithData.value.indexOf(column.element) !== -1)
+    );
+  }
+
+  /**
+   * Marks all columns as invisible to show only the default columns.
+   */
   clearSelection(): void {
     this.columns.forEach((x) => (x.visible = false));
   }