+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
 ## [4.3.2] - 2022-04-01
+### Changed
+- Hide resource table columns without data after loading data.
 ### Fixed
+- The use of a number range filter for the Count column type of
+  the ResearchStudy table has been restored.
 - Replaced incorrect "dateTime" data type for search parameters with "date"
   to use the correct controls.
 - Replaced hyphens in search parameter names with spaces.